// packages/core/src/types/dataGridTypes.d.ts
export type RowId = string | number;
// ...
export interface ISelectionRange {
    startRow: number;
    startCol: number;
    endRow: number;
    endCol: number;
}